pyproject.toml
[tool.poetry]
name = "final-assignment-agent"
version = "0.1.0"
description = "HuggingFace Agents Course Final Assignment - GAIA Benchmark Agent"
authors = ["Sytse van der Schaaf <sytse@svdschaaf.nl>"]
readme = "README.md"
license = "MIT"
homepage = "https://huggingface.co/learn/agents-course"
repository = "https://github.com/sytse06/agents/final-assignment-agent"
documentation = "https://huggingface.co/learn/agents-course"
keywords = ["ai", "agents", "gaia", "benchmark", "huggingface", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.10"

# Core application dependencies
gradio = "^5.13.2"
requests = "^2.32.3"
pandas = "^2.0.0"
python-dotenv = "^1.0.0"

# AI/ML framework - Compatible versions for Gradio 5.x
smolagents = "^1.17.0"
transformers = "^4.40.0"
huggingface-hub = "^0.33.0"  # Required by Gradio 5.x
torch = "^2.1.0"
datasets = "^2.14.0"
evaluate = "^0.4.0" 

# Data processing and file handling
beautifulsoup4 = "^4.12.0"
pillow = "^11.0.0"
openpyxl = "^3.1.0"
pypdf2 = "^3.0.0"
python-docx = "^1.1.0"

# Scientific computing - Compatible with Gradio
numpy = ">=2.0.0,<3.0.0"  # Match Gradio's requirement
matplotlib = "^3.7.0"
scipy = "^1.11.0"

# Audio processing
librosa = "^0.10.0"
soundfile = "^0.12.0"

# Web scraping and API tools
lxml = ">=5.3.0"
selenium = {version = "^4.15.0", optional = true}

# Optional dependencies for enhanced functionality
openai = {version = "^1.0.0", optional = true}
anthropic = {version = "^0.7.0", optional = true}
seaborn = "^0.13.2"
nbstripout = "^0.8.1"
sentence-transformers = "^4.1.0"
weaviate-client = "^4.14.4"
langchain-weaviate = "^0.0.5"
langchain-community = "^0.3.24"
langchain-huggingface = "^0.2.0"
langgraph = "^0.4.8"
langchain-openai = "^0.3.19"
litellm = "^1.72.1"
langchain-ollama = "^0.3.3"
langchain-groq = "^0.3.2"
langchain-google-genai = "^2.1.5"
duckduckgo-search = "^8.0.2"
docling = "^2.36.1"
backoff = "^2.2.1"
markdownify = "^1.1.0"
wikipedia = "^1.4.0"
arxiv = "^2.2.0"
pymupdf = "^1.26.1"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

# Code quality
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.6.0"
pre-commit = "^3.4.0"

# Documentation
mkdocs = "^1.5.0"
mkdocs-material = "^9.4.0"

# Development tools
jupyter = "^1.0.0"
ipykernel = "^6.25.0"
notebook = "^7.0.0"

# Debugging and profiling
ipdb = "^0.13.0"
memory-profiler = "^0.61.0"

[tool.poetry.extras]
selenium = ["selenium"]
openai = ["openai"]
anthropic = ["anthropic"]
all = ["selenium", "openai", "anthropic"]

[tool.poetry.scripts]
fa-agent = "src.agent:main"
fa-test-run = "tests.test_agent:run_interactive_tests"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "smolagents.*",
    "gradio.*",
    "librosa.*",
    "soundfile.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/.*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "E501"]
exclude = [
    ".git",
    "__pycache__",
    "dist",
    "build",
    "*.egg-info",
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
]
per-file-ignores = [
    "__init__.py:F401",
    "tests/*:S101",
]